+++ lib/components/Pane/index.js
@@ -11,7 +11,7 @@ import Button from "../Button";
 import { noop } from "../../utils";
 import { Portal, Backdrop } from "../../atoms";
 import { useOverlay } from "../../hooks/useOverlay";
-import { useOverlayManager } from "../../managers/OverlayManager";
+import { useOverlayManager } from "../../hooks/useOverlayManager";
 
 const SIZES = { small: "small", large: "large" };
