+++ lib/managers/OverlayManager.js
@@ -1,9 +1,13 @@
-import { useEffect } from "react";
-
 class OverlayManager {
   constructor() {
     this.overlays = [];
     this.previouslyFocusedElements = [];
+    this.add = this.add.bind(this);
+    this.remove = this.remove.bind(this);
+    this.isTopOverlay = this.isTopOverlay.bind(this);
+    this.hasOverlays = this.hasOverlays.bind(this);
+    this.getFinalFocusInOverlay = this.getFinalFocusInOverlay.bind(this);
+    this.getTopMostOverlay = this.getTopMostOverlay.bind(this);
   }
 
   add(overlay, elementToFocus) {
@@ -34,15 +38,3 @@ class OverlayManager {
 }
 
 export const manager = new OverlayManager();
-
-export function useOverlayManager(ref, isOpen) {
-  const elementToFocus = document.activeElement;
-  useEffect(() => {
-    if (isOpen) {
-      manager.add(ref, elementToFocus);
-    }
-    return () => {
-      manager.remove(ref);
-    };
-  }, [isOpen, ref]);
-}
