+++ lib/managers/index.js
@@ -1,3 +1,3 @@
 import { manager } from "./OverlayManager";
 
-export { manager as OverlayManager };
+export { manager };
